+++ src/dcos_e2e_cli/dcos_aws/commands/_nodes.py
@@ -8,7 +8,7 @@ import click
 
 from dcos_e2e.node import Node
 
-from ._common import ClusterInstances, InstanceInspectView
+from ._common import ClusterInstances
 
 
 def node_option(command: Callable[..., None]) -> Callable[..., None]:
@@ -37,7 +37,6 @@ def node_option(command: Callable[..., None]) -> Callable[..., None]:
 def get_node(
     cluster_instances: ClusterInstances,
     node_reference: str,
-    aws_region: str,
 ) -> Optional[Node]:
     """
     Get a node from a "reference".
@@ -49,7 +48,6 @@ def get_node(
             * A node's private IP address
             * A node's EC2 instance ID
             * A reference in the format "<role>_<number>"
-        aws_region: The AWS region the cluster is in.
 
     Returns:
         The ``Node`` from the given cluster or ``None`` if there is no such
@@ -62,10 +60,7 @@ def get_node(
     }
 
     for instance in instances:
-        inspect_data = InstanceInspectView(
-            instance=instance,
-            aws_region=aws_region,
-        ).to_dict()
+        inspect_data = cluster_instances.to_dict(node_representation=instance)
         reference = inspect_data['e2e_reference']
         instance_id = inspect_data['ec2_instance_id']
         public_ip_address = inspect_data['public_ip_address']
@@ -88,7 +83,6 @@ def get_nodes(
     node_references: Iterable[str],
     cluster_instances: ClusterInstances,
     inspect_command_name: str,
-    aws_region: str,
 ) -> Set[Node]:
     """
     Get nodes from "reference"s.
@@ -103,7 +97,6 @@ def get_nodes(
             * A reference in the format "<role>_<number>"
         inspect_command_name: The name of an inspect command to use to find
             available references.
-        aws_region: The AWS region the cluster is in.
 
     Returns:
         All ``Node``s from the given cluster which match the references.
@@ -116,7 +109,6 @@ def get_nodes(
         node = get_node(
             cluster_instances=cluster_instances,
             node_reference=node_reference,
-            aws_region=aws_region,
         )
         if node is None:
             message = (