+++ src/dcos_e2e_cli/dcos_aws/commands/send_file.py
@@ -64,7 +64,6 @@ def send_file(
         cluster_instances=cluster_instances,
         node_references=node,
         inspect_command_name=inspect_command_name,
-        aws_region=aws_region,
     )
 
     for host in hosts: